polymarket-bot/requirements.txt
# === Polymarket ===
py-clob-client>=0.34.4
py-order-utils>=0.3.2

# === Weather Data ===
requests>=2.31.0
httpx>=0.27.0

# === CLI ===
typer>=0.9.0
rich>=13.7.0
click>=8.1.7

# === Data & Models ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# === Retry & Resilience ===
tenacity>=8.2.0
backoff>=2.2.0

# === Async ===
aiohttp>=3.9.0
websockets>=12.0

# === Logging ===
structlog>=23.3.0
python-json-logger>=2.0.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# === Type Checking (dev) ===
mypy>=1.7.0
types-requests>=2.31.0
